slack-chatbot/requirements.txt
slackclient==2.9.3
slackeventsapi==3.0.0
Flask==2.0.2
flask-shell-ipython==0.4.1